% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{api_req}
\alias{api_req}
\title{Create GitLab API request}
\usage{
api_req(
  path,
  method = c("GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT",
    "TRACE"),
  url_params = NULL,
  body_json = NULL,
  auto_unbox = TRUE,
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{path}{\code{\link[base:character]{character(1)}}\cr
Path relative to the GitLab API \code{base_url}.}

\item{method}{\code{\link[base:character]{character(1)}}\cr
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods}{HTTP request method}.
One of \code{"GET"}, \code{"CONNECT"}, \code{"DELETE"}, \code{"HEAD"}, \code{"OPTIONS"}, \code{"PATCH"}, \code{"POST"}, \code{"PUT"} or \code{"TRACE"}.}

\item{url_params}{\code{\link[base:list]{list()}}\cr
URL parameters added to the \href{https://en.wikipedia.org/wiki/Query_string}{query string}. Either \code{NULL} to omit or a named list of key-value pairs that
define query parameters. Values must be scalars. To opt out of automatic conversion of non-character values to JSON strings or
\href{https://en.wikipedia.org/wiki/Percent-encoding}{percent-encoding} values, wrap values in \code{\link[=I]{I()}}.}

\item{body_json}{\code{\link[base:list]{list()}}\cr
Data to include as JSON in the HTTP request body. \code{NULL} means an empty body.}

\item{auto_unbox}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to automatically "unbox" length-1 vectors in \code{body_json} to JSON scalars.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of attempts to retry in case of an HTTP error.}
}
\value{
httr2::req_method return
}
\description{
Assembles the HTTP request structure that is common to \emph{all} \href{https://docs.gitlab.com/ee/api/rest/}{GitLab v4 RESTful API} requests performed by this
package.
}
\examples{
# let's call the `/metadata` endpoint (https://docs.gitlab.com/ee/api/metadata.html)
gitlab::api_req(path = "/metadata",
                method = "GET") |>
  httr2::req_perform() |>
  httr2::resp_body_json()
}
\seealso{
Other common functions:
\code{\link{build_url}()},
\code{\link{project}()},
\code{\link{project_default_branch}()}
}
\concept{common}
