% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{dir_exists}
\alias{dir_exists}
\title{Test if directory exists in GitLab repository}
\usage{
dir_exists(
  path,
  id_project = pal::pkg_config_val("id_project"),
  ref = NULL,
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{path}{\code{\link[base:character]{character(1)}}\cr
File path, relative to the repository root.}

\item{id_project}{\code{\link[base:integer]{integer(1)}}\cr
GitLab project identifier.}

\item{ref}{\code{\link[base:character]{character(1)}}\cr
\href{https://git-scm.com/docs/revisions#_specifying_revisions}{Git revision expression} matching the desired Git tree object, e.g. a ref name (branch, tag,
etc.), a commit identifier, or another symbolic reference like \code{"HEAD~10"}. Omitted if \code{NULL}, otherwise set as a URL parameter in the API request. Note
that the GitLab API doesn't support every type of revision expression.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of attempts to retry in case of an HTTP error.}
}
\value{
A logical scalar.
}
\description{
Tests whether or not a directory (Git tree object) exists in a GitLab repository.
}
\details{
\code{dir_exists()} makes a light-weight HTTP HEAD request only and properly handles HTTP 404 Not Found errors.
}
\examples{
gitlab::dir_exists(path = "docs",
                   id_project = 64767928)

gitlab::dir_exists(path = "no/no/nope",
                   id_project = 64767928)
}
\seealso{
Other functions to manage directories on GitLab:
\code{\link{dir_delete}()},
\code{\link{dir_ls}()}
}
\concept{dirs}
