% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{file_meta}
\alias{file_meta}
\title{GitLab file metadata}
\usage{
file_meta(
  path,
  attribute = c("file_name", "file_path", "size", "encoding", "content_sha256", "ref",
    "blob_id", "commit_id", "last_commit_id", "execute_filemode"),
  id_proj,
  ref = NULL,
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{path}{\code{\link[base:character]{character(1)}}\cr
File path, relative to the repository root.}

\item{attribute}{\code{\link[base:character]{character(1)}}\cr
Name of the metadata attribute to return.}

\item{id_proj}{\code{\link[base:integer]{integer(1)}}\cr
GitLab project identifier.}

\item{ref}{\code{\link[base:character]{character(1)}}\cr
\href{https://git-scm.com/docs/revisions#_specifying_revisions}{Git revision expression} matching the desired Git tree object, e.g. a ref name (branch, tag,
etc.), a commit identifier, or another symbolic reference like \code{"HEAD~10"}. Note that the GitLab API doesn't support every type of revision expression.
Defaults to \code{"HEAD"} if \code{NULL}.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of attempts to retry in case of an HTTP error.}
}
\value{
An integer scalar if \code{attribute = size}, a logical scalar if \code{attribute = execute_filemode}, or a character scalar in all other cases.
}
\description{
Retrieves a file's metadata via the
\href{https://docs.gitlab.com/ee/api/repository_files.html#get-file-metadata-only-1}{\code{HEAD /projects/{id}/repository/files/{path}}} endpoint of GitLab's v4
RESTful API.
}
\examples{
gitlab::file_meta(path = "README.md",
                  attribute = "execute_filemode",
                  id_proj = 64767928)
}
\seealso{
Other file management functions:
\code{\link{file_content}()},
\code{\link{file_create}()},
\code{\link{file_delete}()},
\code{\link{file_exists}()},
\code{\link{file_full}()},
\code{\link{file_req}()},
\code{\link{file_update}()},
\code{\link{file_write}()}
}
\concept{files}
