% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{file_req}
\alias{file_req}
\title{Create GitLab file management API request}
\usage{
file_req(
  method,
  path,
  id_project = pal::pkg_config_val("id_project"),
  ref = NULL,
  body_json = NULL,
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{method}{\code{\link[base:character]{character(1)}}\cr
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods}{HTTP request method}.
One of \code{"GET"}, \code{"CONNECT"}, \code{"DELETE"}, \code{"HEAD"}, \code{"OPTIONS"}, \code{"PATCH"}, \code{"POST"}, \code{"PUT"} or \code{"TRACE"}.}

\item{path}{\code{\link[base:character]{character(1)}}\cr
File path, relative to the repository root.}

\item{id_project}{\code{\link[base:integer]{integer(1)}}\cr
GitLab project identifier.}

\item{ref}{\code{\link[base:character]{character(1)}}\cr
\href{https://git-scm.com/docs/revisions#_specifying_revisions}{Git revision expression} matching the desired Git tree object, e.g. a ref name (branch, tag,
etc.), a commit identifier, or another symbolic reference like \code{"HEAD~10"}. Omitted if \code{NULL}, otherwise set as a URL parameter in the API request. Note
that the GitLab API doesn't support every type of revision expression.}

\item{body_json}{\code{\link[base:list]{list()}}\cr
Data to include as JSON in the HTTP request body. \code{NULL} means an empty body.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of attempts to retry in case of an HTTP error.}
}
\value{
httr2::req_method return
}
\description{
Assembles the HTTP request structure that is common to \emph{all} GitLab v4 RESTful API requests to manage files performed by this package.
}
\examples{
gitlab:::file_req(method = "HEAD",
                  path = "README.md",
                  id_project = 64767928,
                  ref = "HEAD")
}
\seealso{
Other functions to manage files on GitLab:
\code{\link{commit_files}()},
\code{\link{file_commit_action}()},
\code{\link{file_content}()},
\code{\link{file_create}()},
\code{\link{file_delete}()},
\code{\link{file_exists}()},
\code{\link{file_full}()},
\code{\link{file_meta}()},
\code{\link{file_update}()},
\code{\link{file_write}()},
\code{\link{files_delete}()}
}
\concept{files}
\keyword{internal}
