% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-package.R
\docType{package}
\name{gitlab-package}
\alias{gitlab}
\alias{gitlab-package}
\title{gitlab: Use GitLab's v4 RESTful API}
\description{
Allows to conveniently work with GitLab's v4 RESTful API from R. GitLab (https://gitlab.com) is an open-source DevOps software suite that i.a. includes a powerful Git forge.
}
\section{Package configuration}{


Some of gitlab's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Base URL of the GitLab v4 RESTful API server to be accessed when \code{base_url} is not explicitly provided to the functions of this package. \tab \code{gitlab.base_url} \tab \code{R_GITLAB_BASE_URL} \tab \code{"https://gitlab.com/api/v4"} \cr
   \href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication when \code{token} isnot explicitly provided to the functions of this package. A personal access token for gitlab.com can be created under\href{https://gitlab.com/-/user_settings/personal_access_tokens}{this link}. \tab \code{gitlab.token} \tab \code{R_GITLAB_TOKEN} \tab \code{Sys.getenv("GITLAB_COM_TOKEN")} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/gitlab}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/gitlab/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
