% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{project}
\alias{project}
\title{GitLab project metadata}
\usage{
project(
  id_project = pal::pkg_config_val("id_project"),
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{id_project}{\code{\link[base:integer]{integer(1)}}\cr
GitLab project identifier.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of request attempts in case of an HTTP error. An integerish scalar. Retries are performed using exponential backoff and jitter, see \code{\link[httr2:req_retry]{httr2::req_retry()}} for details.}
}
\value{
A list.
}
\description{
Retrieves a GitLab project's metadata via the \href{https://docs.gitlab.com/ee/api/projects.html#get-a-single-project}{\code{GET /projects/{id_project}}} endpoint of
GitLab's v4 RESTful API.
}
\examples{
gitlab::project(id_project = 64767928) |> names()
}
\seealso{
Other functions to manage GitLab projects:
\code{\link{project_default_branch}()}
}
\concept{projects}
