% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.gen.R
\name{project_default_branch}
\alias{project_default_branch}
\title{GitLab project default branch}
\usage{
project_default_branch(
  id_proj,
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)
}
\arguments{
\item{id_proj}{\code{\link[base:integer]{integer(1)}}\cr
GitLab project identifier.}

\item{base_url}{\code{\link[base:character]{character(1)}}\cr
Base URL to the GitLab v4 RESTful API root.}

\item{token}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication. If \code{NULL}, the API request is made unauthenticated
which results in \emph{HTTP 404 Not Found} and other errors for non-public resources.}

\item{max_tries}{\code{\link[base:integer]{integer(1)}}\cr
Maximum number of attempts to retry in case of an HTTP error.}
}
\value{
A character scalar.
}
\description{
Determines a GitLab project's default Git branch name.
}
\examples{
gitlab::project_default_branch(id_proj = 64767928)
}
\seealso{
Other common functions:
\code{\link{api_req}()},
\code{\link{build_url}()},
\code{\link{project}()}
}
\concept{common}
